Begin cleaning up lints; untangle dependency cycle
