+++ src/cli/config.ts
@@ -10,6 +10,8 @@ import { BabblerBaseApp } from "../apps/babbler/base";
 import { IWritableToken } from "../token";
 import { configInPath, readConfig } from "./commands/config";
 
+export { getAppConstructors } from "./getAppConstructors";
+
 const debug = _debug("babbling:config");
 
 export const DEFAULT_CONFIG_PATH = pathlib.join(
@@ -17,57 +19,7 @@ export const DEFAULT_CONFIG_PATH = pathlib.join(
     ".config/babbling/auto-config.json",
 );
 
-export async function* getAppConstructors(): AsyncIterable<IAppConstructor<any, IApp>> {
-    const allExports = require("../index");
-    for (const name of Object.keys(allExports)) {
-        if (
-            name.endsWith("App")
-            && !name.endsWith("BaseApp")
-            && allExports[name] instanceof Function
-        ) {
-            yield allExports[name];
-        }
-    }
-}
-
-export async function* importConfig(configPath?: string) {
-    const actualPath = configPath || DEFAULT_CONFIG_PATH;
-    const config = await readConfig(actualPath);
-
-    yield* importConfigFromJson(actualPath, config);
-}
-
 type ConfigPair<TOpts extends Opts> = [IPlayerEnabledConstructor<TOpts, IApp>, Opts];
-
-export async function* importConfigFromJson(
-    configPath: string,
-    config: any,
-) {
-    for await (const app of getAppConstructors()) {
-        const appConfig = config[app.name];
-        if (!appConfig) continue;
-
-        if (
-            app.prototype instanceof BabblerBaseApp
-            && config.babbler
-            && !appConfig.appId
-        ) {
-            appConfig.appId = config.babbler;
-        }
-
-        if (app.tokenConfigKeys) {
-            for (const k of app.tokenConfigKeys) {
-                appConfig[k] = new AppConfigToken(
-                    configPath,
-                    [app.name, k],
-                    appConfig[k],
-                );
-            }
-        }
-
-        yield [app, appConfig] as ConfigPair<any>;
-    }
-}
 
 class AppConfigToken implements IWritableToken {
     constructor(
@@ -99,4 +51,41 @@ class AppConfigToken implements IWritableToken {
     public toString(): string {
         return this.value;
     }
+}
+
+export async function* importConfigFromJson(
+    configPath: string,
+    config: any,
+) {
+    for await (const app of getAppConstructors()) {
+        const appConfig = config[app.name];
+        if (!appConfig) continue;
+
+        if (
+            app.prototype instanceof BabblerBaseApp
+            && config.babbler
+            && !appConfig.appId
+        ) {
+            appConfig.appId = config.babbler;
+        }
+
+        if (app.tokenConfigKeys) {
+            for (const k of app.tokenConfigKeys) {
+                appConfig[k] = new AppConfigToken(
+                    configPath,
+                    [app.name, k],
+                    appConfig[k],
+                );
+            }
+        }
+
+        yield [app, appConfig] as ConfigPair<any>;
+    }
+}
+
+export async function* importConfig(configPath?: string) {
+    const actualPath = configPath || DEFAULT_CONFIG_PATH;
+    const config = await readConfig(actualPath);
+
+    yield* importConfigFromJson(actualPath, config);
 }
